try to compare version

fix conflict


remove unuse change
